src/requirements.txt
# OpenAI API
openai>=1.3.0

# Document parsing
python-docx>=1.1.0
python-pptx>=0.6.23

# Environment variables
python-dotenv>=1.0.0

# Optional: For development
pytest>=7.4.0
pytest-cov>=4.1.0

